service-2/requirements.txt
asgiref==3.6.0
Django==4.2.1
django-configurations==2.4.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Flask==2.2.3
grip==4.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
Markdown==3.4.3
MarkupSafe==2.1.2
path-and-address==2.0.1
psycopg2-binary==2.9.6
Pygments==2.14.0
PyJWT==2.7.0
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
Werkzeug==2.2.3
whitenoise==6.4.0
uwsgi
